requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
drf-yasg-stubs==0.1.4
inflection==0.5.1
Markdown==3.5.2
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
uritemplate==4.1.1
